Refactor postal service stats handling
